test: add minimal agent stubs for tests
